move main command functionality into the sub-command
