+++ src/AzureAuth.Test/AzureAuth.Test.csproj
@@ -45,4 +45,8 @@
     <ProjectReference Include="..\AzureAuth\AzureAuth.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Commands\" />
+  </ItemGroup>
+
 </Project>
