+++ src/stylecop/GlobalSuppressions.Test.cs
@@ -5,3 +5,4 @@ using System.Diagnostics.CodeAnalysis;
 
 [assembly: SuppressMessage("StyleCop.CSharp.SpecialRules", "SA0001: XML comment analysis is disabled due to project configuration", Justification = "Not needed in test project.")]
 [assembly: SuppressMessage("StyleCop.CSharp.OrderingRules", "SA1600: Elements should be documented", Justification = "Not needed in test project.")]
+[assembly: SuppressMessage("StyleCop.CSharp.ReadabilityRules", "SA1101:Prefix local calls with this", Justification = "test file")]
